radium_orchestrator::orchestration

Module git_extended_tools

Extended Git and code search tools

This module provides advanced Git and code search capabilities:

find_references: Find all references to a symbol using ripgrep

git_blame: Show git blame for a file

git_show: Show git commit details

Functions

create_find_references_tool

Create the find_references tool

create_git_blame_tool

Create the git_blame tool

create_git_diff_tool

Create the git_diff tool

create_git_extended_tools

Create all extended Git and search tools

create_git_log_tool

Create the git_log tool

create_git_show_tool

Create the git_show tool

create_git_status_tool

Create the git_status tool

find_symbol_references

Find all references to a symbol using internal search

git_blame

Execute git blame on a file

git_show

Execute git show for a commit